Add generation of schema via jsonschema for query validation
Adds the generation of json-schema schema. This should enable both the
frontend and backend to validate the query data.

At some point I may even be able to utilize the schema for the
building of the frontend options, but at this time the recursive
nature of the schema makes that implementation non-trivial.