Provides a schema for the query data filtering options
Provides a schema for the query data filtering options to be
represented in JSON so that the frontend can display the options to
the user.

Ideally this is replaced by the use of the json-schema in the future,
but at this time, this solution provides an easy way for the frontend
to grapple with the options.