Remove custom JSON encoder
Removes the custom JSON encoder that was added to account for the use
of the set type. It turns out that `jsonschema` doesn't like the use
of the set type either. When attempting to validate a schema
containing sets we run into an exception.

Easiest way to fix this is just to use lists instead. I'm still using
sets in the lookups generator. I'm not entirely sure that's necessary,
but won't hurt too much either.