Implement explicit filter definition; removes meta generation
Removes meta class filter generation in favor of explicit filter
definitions through options. Objects are being used to provide extra
details that are not going to be easily conveyed through dicts and
strings in a meta class. At some point I'll bring back the feature to
define filters in the meta class, but for now it only gets in the way.