Raise any metadata exceptions with more detail from the type
The exceptions that come out of the Metadata class are fairly
difficult to inspect through the traceback. This is due to when the
code is executed. In order to give the developer some kind of cue,
catch the exception in the FieldSet type where we have access to the
class name, module, etc.