Ensure query data uses only lists instead of tuples
Moving away from tuples to ensure the json encoding and decoding cycle
results in the same data structures. This also helps with checking for
individual filter quality.

This change was prompted from the need to translate form based
filtering to the query structure. Part of that translation is to check
if the form key and value are already present in the query data. It
could be in tuple form if the query data was just translated or it
might be in list format if translated from json. This makes it
challenging to assert equality with an existing filter.