Rework creation of Q instance from query data
Reworked backend creation of Q instance so that the creation of this
object utilizes the filter objects definition to translate the value.

It perhaps isn't apparent why this is necessary. At this time we
simply pass through the value as given. However, other lookup types
may need to translate the data to something more native or something
completely different. For example, the `range` lookup takes a tuple of
two datetimes. With this change we can code the filter's lookup to
split the value and return the tuple of two datetime objects.

Another example might be sticky filters, which are present in the
query regardless of user inclusion but disappear if the user has
explicitly set the filter to something that equates to "any". In this
case our filter lookup will return `None` indicating the filter won't
be present in the query.

These example cases are yet to be implemented, but the belief is that
this move to enabling the filter lookups to translate the value will
be helpful in achieving them.