Adds a sticky filter that provides a default query
A sticky filter is type of filter that always has a query except when
the user specifically selects to disable or remove it. This feature
enables default filters to be applied upon first use.

The usecase that drove this implementation was for a page that lists
objects from a model that has a `type` field. This filter field upon
first visit is assumed to be set to 'manual'. The user is provided
with the option to change the value to 'bulk' or 'any'. The 'any'
value eliminates the filter from the query. It's a sort of inverted
filter, where instead of starting with no filters applied, we start
with an applied filter.