Include a default lookup in criteria construction for query data
Is it absolutely necessary to provide a lookup value in the query data
for a condition? No, but I'd like to say "explicit over implicit" as a
general rule. But also requiring the lookup could help to keep JSON
Schema validation simple and correct.

Is it possible to translate fields without lookups to filters? That
is, in a traditional sense where a `Form` is being adapted to a
`FilterSet`'s query data, can a field without a lookup be translated?
Yes. It should be assumed that the `exact` lookup is in use. In fact,
Django makes this decision for us because the default lookup for any
field filter is `exact`.