Require an initial QuerySet instance be given to FilterSet
Change to require a QuerySet instance be given to
FilterSet.filter_queryset.

Made to address need to make FilterSet non-model specific
FilterSet. In this scenario the FilterSet is not tied to Model, but
instead can be generically used. This is in an effort to allow the
FilterSet to handle polymorphic models.