Enable the use of a custom translator
Allow the developer to define a custom translator function, which can
be used to provide more complex filtering criteria.

This feature enables side effects to translate to  filter
criteria. Some examples uses of this include:
- psuedo fields
- made up datapoints
- api data points
- static data
- cross database queries

The translation testcase illustrates a usecase where the model does
not have a `continent` field, but is able to translate the continent
into country codes we can use to filter the model. The underlying
logic for this is hardcoded, but could anything from a api call to a
cross database query.

Admittedly, the solution provided for this feature is not the best. I
suspect future iterations will clean up the interface for
this. Perhaps I'll look at special methods on the FilterSet, similar
to the django form's field cleaning methods.