Return Q objects from Filters
Return Q objects from Filter instead of Q arguments. This allows the
developer to return criteria in a single object and allows the
Filterset to intelegently union it with the sibling or parent
criteria.

Renamed some methods in the process of making this change. Translate
has become transmute. Some methods now have simplified names to
encourage a more memorable API (mostly for me).

And example where the returning of a Q object with two bits of
criteria is a date range. From the interface this might be a "between"
lookup that one the backend us up as a 'gte' and 'lte' lookup on a
date field.