Allow the generation of the options schema without a model field
Allows the generation of the options schema by a Filter without having
been given a Model Field. This better supports the idea that filters
do not need to be tied to a particular field. Furthermore, because the
Filter is bound on FilterSet instantiation, we're able to discover the
field when needed.