Allows custom filter transmute definitions on the FilterSet
These adjustments to the transmute logic allow for the custom
`transmute_<filter>[__<lookup...>]` methods on the FilterSet
class. This gives the developer the ability to override the
transmutation logic within the Filter or Filter Lookup classes without
needing to create custom classes.

The transmutation logic is now given a full set of contextual
infomation. This was necessary to ensure transmute methods on the
FilterSet as well as in the Lookups (opposite end of the resolution
path) all have the same contextual information.

This change was brought about by the need to address a type of filter
that does not have a single lookup or a single value. The case in
question is a filter for a date range. The lookups will be `gte` and
`lte` and the value in the criteria will be an array of values or JSON
object. The Filter class up to this point has been singularly focused
on a single model field and lookup solution.

With this adjustment, it should be possible to built another type of
Filter that is able to ignore the lookup concept entirely and/or defer
to the Lookup for precise tranmutation logic.