Redefine transmute signature for clarity, expansion and typing
Rethinks the arguments to transmute methods where the keyword argument
approach does a poor job of letting the developer know what arguments
are expected and of what type. It also helps the python typing.

This works towards the removal of FilterSet from the Filter's bind
method and Filter itself. The Filter will instead have access to the
FilterSet through context info where needed.