Move Filter name binding back to FilterSet metaclass
Moves the Filter's name binding procedure back to the FilterSet
metaclass. Assigning the name to a Filter was originally in the
FilterSet metaclass (i.e. FilterSetType). The binding of the name was
moved to instantiation because there was a desire to also bind the
FilterSet instance. The need for the Filter to have access to the
FilterSet instance has been resolved through the passing around of
context info where needed. Without the need to bind the FilterSet
instance, the attribute name binding can move back to the metaclass.