Fix TypeError: unhashable type: 'list' for lookup name
The Lookup name cannot be a `list` if we intend to use the name as a
dictionary key. Originally, the thought was to separate the lookups,
because they are distinct values each that tie to a specific field or
lookup code. However, the `list` type cannot be used as a dictionary
key. In hindsight it's likely easier for the developer to type
the lookup in the representation they know already. Plus, we always
have the ability to split the name into a list where needed.