Correct help_text to use the correct field
The help_text was using the last remnant of the loop that resolves the
filter and lookup to a field. This means the help text will be
incorrect for any relational field lookups.

Also, the `Field.help_text` property is lazy. This causes an issue
when we JSON encode the schema. So make the value active by casting it
to a basic JSON encodable type.