Adds functionality to generate filters from a model
Provides a function to generate filters from a model's fields. This
was modeled after `django.forms.models.fields_for_model`, which
produces form fields for a model.

Later changes to the Filter metaclass will use the `filters_for_model`
to produce filters for a model's fields. This shortcuts the need to
define each field as a filter on the FilterSet.

The default factory functions for filter and lookup are exposed at
this time. The though is that a developer could define a
`filter_factory_callback` that could defer to the default behavior if
necesary.

Hash methods have been added to the Filter and Lookup classes to allow
the tests to easily check for expectation equality.