Create Filters from Meta defined fields and lookups
Allows for filters to be generated for a model's fields using the Meta
class when declaring the FilterSet. This works similar to how Django
Model Forms work, in that one can specify the model and fields. In
this case we ask for the field and lookups to use.

This takes out some of the overhead around defining FilterSets. In
simple cases, the developer can now define the FilterSet using a
dictionary of fields mapped to a list of lookups.