Add app configuration for settings lookup
Adds a means of obtaining app settings for this app.

This includes a mapping of lookup names to labels for the declarative
generation of filters. (A followup commit will connect this
configutation.) In this case a default setting is provided, but can be
overridden by the developer through django project settings.