Provide flat Form representation of FilterSet's first level data
Provides a Form class that enables submission of filters via a form
that can update the FilterSet instance.

The creation of the flat Form class is generated using the
FilterSet. This utilizes special `as_form_fields` methods on the
Filter and Lookup instances. Ideally the design wouldn't put the form
implementation directly in these classes, but it's the easiest path
forward at the moment; also it follows closely with what Django does
with Model Fields to Form Fields.