Provide a flag for when the flat form is enabled
The form will not accurately represent the user's intentions when they
use operators other than 'and' or are using nested filters.
The `is_enabled` property is intended to be used by the developer to
show or hide the form.

Generally speaking, the form is disabled when the user has entered
__advanced mode__ and a flat form interface will no longer accurately
represent the filtering strategy.