Fix choice lookup form field creation to use model field choices
A Model Field can directly define its choices or generate them through
a relation. The additions here enable the use of the choices in those
scenarios. This follows similar logic found in the Field.formfield
method and RelatedField.formfield method. The result is a form field
with the correct choices within it.